+++ core/Cargo.toml
@@ -20,7 +20,6 @@ ckb-util = { path = "../util" }
 fnv = "1.0.3"
 ckb-merkle-tree = {path = "../util/merkle-tree"}
 faster-hex = "0.3"
-serde_bytes = "0.11.1"
 bytes = { version="0.4.12", features = ["serde"] }
 log = "0.4"
 failure = "0.1.5"
