+++ verification/src/lib.rs
@@ -1,12 +1,15 @@
 mod block_verifier;
+mod contextual_block_verifier;
 mod error;
 mod header_verifier;
 mod transaction_verifier;
+mod uncles_verifier;
 
 #[cfg(test)]
 mod tests;
 
-pub use crate::block_verifier::{BlockVerifier, HeaderResolverWrapper, TransactionsVerifier};
+pub use crate::block_verifier::{BlockVerifier, HeaderResolverWrapper};
+pub use crate::contextual_block_verifier::ContextualBlockVerifier;
 pub use crate::error::{Error, TransactionError};
 pub use crate::header_verifier::{HeaderResolver, HeaderVerifier};
 pub use crate::transaction_verifier::{
